refactor: Web Controller refactored the web controllers to be inside of a 'Web' folder for cleaner, neater code structure. reconnected the 'use' statements in the web routes file.
