+++ app/Http/Controllers/Web/ClassSessionController.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Web;
 
 use App\Models\ClassSession;
 use App\Models\Cluster;
 use App\Models\User;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
 
 class ClassSessionController extends Controller
 {
